backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose
passlib[bcrypt]
minio
pydantic
sqlalchemy
pydantic[email]